backend/requirements.txt
sanic==21.9.3
sanic-cors==1.0.1
mysql-connector-python==8.0.28